+++ src/components/Common/Header.tsx
@@ -5,11 +5,31 @@ import CartIcon from "../../assets/svg/cart-icon.svg";
 import UserIcon from "../../assets/svg/user-icon.svg";
 import SignupIcon from "../../assets/svg/signup-icon.svg";
 import { useState } from "react";
+import { baseInstance } from "../../hooks/useAxios";
 
 const Header = () => {
   const [loggedin, setLoggedin] = useState(false);
   const handleLogin = () => {
     setLoggedin((prev) => !prev);
+    fetchData();
+  };
+
+  const fetchData = async () => {
+    try {
+      const response = await baseInstance.post("/members/login", {
+        email: "test@nam.com",
+        password: "1234",
+      });
+
+      const accessToken = response.headers["access_token"];
+      const refreshToken = response.headers["refresh_token"];
+
+      localStorage.setItem("access_token", accessToken);
+      localStorage.setItem("refresh_token", refreshToken);
+      console.log(response.data);
+    } catch (error) {
+      console.log(error);
+    }
   };
 
   return (
